Support Bash highlighting
When copying a section of Bash script that doesn't include the
`#/bin/bash` shebang line, vim assumes that the copied text is ksh and
fails to properly highly bashism like `"${file/#*./.}"`. The vim
variable `b:is_bash` is used to set the type to bash.

This checks if `b:is_bash` is defined for the current buffer and if so,
sets it in the new buffer. It's important that this happen before
setting the filetype, although I didn't dig into why that is so.